refactor: 기존 db, sdk 사용하도록 변경
