+++ loan-mate/build.gradle
@@ -16,7 +16,15 @@ java {
 }
 
 repositories {
+	mavenLocal()
 	mavenCentral()
+	maven {
+		url = uri("https://maven.pkg.github.com/fisa-5th-csd-final/Core-Bank-Persistence")
+		credentials {
+			username = System.getenv("GITHUB_USERNAME")
+			password = System.getenv("GITHUB_TOKEN")
+		}
+	}
 }
 
 dependencies {
@@ -27,8 +35,10 @@ dependencies {
 	annotationProcessor 'org.projectlombok:lombok'
 	testImplementation 'org.springframework.boot:spring-boot-starter-test'
 	testRuntimeOnly 'org.junit.platform:junit-platform-launcher'
-    // WebClient 사용을 위한 webflux 관련 의존성 추가
-    implementation 'org.springframework.boot:spring-boot-starter-webflux'
+	// WebClient 사용을 위한 webflux 관련 의존성 추가
+	implementation 'org.springframework.boot:spring-boot-starter-webflux'
+	// corebanking sdk
+	implementation "com.fisa.bank:persistence:0.0.1-SNAPSHOT"
 }
 
 // spotless
