+++ loan-mate/src/main/java/com/fisa/bank/loan/application/model/Loan.java
@@ -1,0 +1,15 @@
+package com.fisa.bank.loan.application.model;
+
+import lombok.Getter;
+import lombok.RequiredArgsConstructor;
+
+import com.fisa.bank.loan.persistence.enums.RiskLevel;
+
+@Getter
+@RequiredArgsConstructor
+public class Loan {
+  private final Long loanId;
+  private final String loanName;
+  private RiskLevel riskLevel;
+  private LoanDetail loanDetail;
+}
