+++ loan-mate/src/main/java/com/fisa/bank/loan/application/repository/LoanRespository.java
@@ -1,10 +1,11 @@
 package com.fisa.bank.loan.application.repository;
 
-import com.fisa.bank.loan.application.domain.Loan;
-
 import java.util.List;
+
+import com.fisa.bank.loan.application.model.Loan;
+import com.fisa.bank.persistence.user.entity.id.UserId;
 
 public interface LoanRespository {
 
-    List<Loan> getLoans(Long userId);
+  List<Loan> getLoans(UserId userId);
 }
